Update main.go
Need to add in error checking. Each Dell service tag should be checked. Might be able to regex even with the | but I'll need to think about how. for now splitting each at the | and then checking and dropping ones that dont match may be easier. Then rebuild with the separator? Just a thought.