Created Web Interface
Created web interface and fiddled with the raw text Dell returns to
force objects which can be returned as either an array or a single item
to always be an array. Also, due to some numbers being 0-prefixed,
sometimes they would come down as strings, and other times as floats.
Used regular expressions to replace any bare numbers with quotes to
force them to act as a string for consistency.